requirements.txt
websockets
bleak
pillow
crccheck
dotenv
paho
